+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Common style tags feature: Add comma-separated tags to all image prompts with configurable prefix/suffix position and automatic deduplication
+
 ## [1.0.0] - 2025-10-09
 
 Initial release of SillyTavern Auto Illustrator extension.